# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Geocoding SaaS API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface GeocodeBatchRequest {
  /** Gaode Maps Web Service API Key */
  apiKey: string;
  /** List of addresses to geocode */
  addresses: string[];
}

export type GeocodeResultStatus =
  (typeof GeocodeResultStatus)[keyof typeof GeocodeResultStatus];

export const GeocodeResultStatus = {
  success: "success",
  failed: "failed",
} as const;

export interface GeocodeResult {
  /** Original address */
  address: string;
  /** Longitude */
  lng?: string;
  /** Latitude */
  lat?: string;
  /** Formatted address returned by API */
  formattedAddress?: string;
  status: GeocodeResultStatus;
  /** Error message if failed */
  error?: string;
}

export interface GeocodeBatchResponse {
  results: GeocodeResult[];
  total: number;
  success: number;
  failed: number;
}

export interface ErrorResponse {
  error: string;
  message: string;
}