# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Geocoding SaaS API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Geocode a list of addresses using Gaode Maps API
 * @summary Batch geocode addresses
 */
export const GeocodeBatchBody = zod.object({
  apiKey: zod.string().describe("Gaode Maps Web Service API Key"),
  addresses: zod.array(zod.string()).describe("List of addresses to geocode"),
});

export const GeocodeBatchResponse = zod.object({
  results: zod.array(
    zod.object({
      address: zod.string().describe("Original address"),
      lng: zod.string().optional().describe("Longitude"),
      lat: zod.string().optional().describe("Latitude"),
      formattedAddress: zod
        .string()
        .optional()
        .describe("Formatted address returned by API"),
      status: zod.enum(["success", "failed"]),
      error: zod.string().optional().describe("Error message if failed"),
    }),
  ),
  total: zod.number(),
  success: zod.number(),
  failed: zod.number(),
});