# lib/api-zod/src/generated/types/geocodeBatchRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Geocoding SaaS API
 * OpenAPI spec version: 0.1.0
 */

export interface GeocodeBatchRequest {
  /** Gaode Maps Web Service API Key */
  apiKey: string;
  /** List of addresses to geocode */
  addresses: string[];
}