# lib/api-zod/src/generated/types/geocodeBatchResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Geocoding SaaS API
 * OpenAPI spec version: 0.1.0
 */
import type { GeocodeResult } from "./geocodeResult";

export interface GeocodeBatchResponse {
  results: GeocodeResult[];
  total: number;
  success: number;
  failed: number;
}