# lib/api-zod/src/generated/types/geocodeResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Geocoding SaaS API
 * OpenAPI spec version: 0.1.0
 */
import type { GeocodeResultStatus } from "./geocodeResultStatus";

export interface GeocodeResult {
  /** Original address */
  address: string;
  /** Longitude */
  lng?: string;
  /** Latitude */
  lat?: string;
  /** Formatted address returned by API */
  formattedAddress?: string;
  status: GeocodeResultStatus;
  /** Error message if failed */
  error?: string;
}